Passing unsigned values to `EXTEND_SIGN`

+++ src/vmem.c
@@ -29,7 +29,7 @@ static uintptr_t parse_source_address(PackedInstruction *instr_ptr, HostThreadDa
 		case 0b0000011: {
 				int addr_reg = unpacked.rs1;
 				int data_reg = unpacked.rd;
-				int imm = EXTEND_SIGN((unpacked.funct7 << 5) | unpacked.rs2, 12);
+				int imm = EXTEND_SIGN((unsigned) (unpacked.funct7 << 5) | unpacked.rs2, 12);
 				uintptr_t reg_value = addr_reg ? ctx->active_regs.x_plus_one[addr_reg - 1] : 0;
 				uintptr_t addr = reg_value + imm;
 				if (reg_out) {
@@ -162,7 +162,7 @@ static uintptr_t parse_destination_address(PackedInstruction *instr_ptr, HostThr
 		case 0b0100011: {
 				int addr_reg = unpacked.rs1;
 				int data_reg = unpacked.rs2;
-				int imm = EXTEND_SIGN((unpacked.funct7 << 5) | unpacked.rd, 12);
+				int imm = EXTEND_SIGN((unsigned) (unpacked.funct7 << 5) | unpacked.rd, 12);
 				uintptr_t reg_value = addr_reg ? ctx->active_regs.x_plus_one[addr_reg - 1] : 0;
 				uintptr_t addr = reg_value + imm;
 				if (reg_out) {
